refactored config panel into module
