+++ src/modules/ConfigPanel/ConfigPanel.module.css
